Move encrypted DNS state to crate shared between iOS and everyeone else
